refactor: change authpage folder location
