+++ src/pages/AuthPage/index.js
